\name{alffinder}
\alias{alffinder}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Parameter alpha in the fitting with a Gamma distribution
}
\description{
Finds parameter alpha in the case of fitting with a Gamma distribution
}
\usage{
alffinder(res, slope, ro, c, z, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
intercept of first moments acorss area
}
  \item{slope}{
slope of first moments acorss area
}
  \item{ro}{
slope of number of individuals across area
}
  \item{c}{
parameter c in the power-law species area relationship
}
  \item{z}{
parameter z in the power-law species area relationship
}
  \item{A}{
total area of consideration
}
}