\name{BellB}
\alias{BellB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bell function
}
\description{
Returns the values of Bell function
}
\usage{
BellB(n, k, x) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
total elements of the sets
}
  \item{k}{
number of partitions
}
  \item{x}{
specific element(s) of the set
}
}