\name{bet.f}
\alias{bet.f}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
bet.f
}
\description{
returns beta(n,N) as a part of finding values of Tchebychev polynomials.
}
\usage{
bet.f(n, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
order of polynomial
}
  \item{N}{
number of bins
}
}
