\name{calcalpha}
\alias{calcalpha}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Logseries parameter
}
\description{
Finds the parameter alpha in the case of fitting with a log-series distribution
}
\usage{
calcalpha(n.orig, s.orig)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n.orig}{
number of individuals
}
  \item{s.orig}{
number of species
}
}