\name{cmoment}
\alias{cmoment}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
moment calculator, first step in the Log-Normal, values need to be transformed
}
\description{
First step for calculating moments of log-normal, will be used in another function for transformation.
}
\usage{
cmoment(mu, sig, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mu}{
the parameter mu in the log-normal
}
  \item{sig}{
the parameter sigma in the log-normal
}
  \item{k}{
maximum order of moments
}
}
