\name{cmul}
\alias{cmul}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
symbolic constant multiplication
}
\description{
symbolic multiplication with a constant
}
\usage{
cmul(u, c)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
  \item{c}{
a constant
}
}