\name{Coef.Tcheb.pol.2nd}
\alias{Coef.Tcheb.pol.2nd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coefficients of the Tchebychev polynomials up to a certain degree
}
\description{
Coefficients of the Tchebychev polynomials up to a certain degree as a matrix
}
\usage{
Coef.Tcheb.pol.2nd(N, no.m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Number of bins
}
  \item{no.m}{
maximum order of polynomials
}
}