\name{data_read}
\alias{data_read}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get values from keyboard for future use.
}
\description{
Take a value from keyboard for future use. It is a part of LLfun which is to find the best fitting distribution for a smaller sub-areas.
}
\usage{
data_read()
}
