\name{expnd}
\alias{expnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
symbolic expansion
}
\description{
expand symbolic polynomials to higher degree
}
\usage{
expnd(u, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
  \item{n}{
numeric value that we need to expand the polynomial.
}
}
