\name{extpfun}
\alias{extpfun}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
extrapolation of moments across scales
}
\description{
extrapolate moments.
}
\usage{
extpfun(dat, m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
list of values (the list must have almost linear pattern).
}
  \item{m}{
the corresponding numeric index in the list that we try to extrapolate to.
}
}