\name{extrapolation.dist}
\alias{extrapolation.dist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
extrapolate distribution moments across scales
}
\description{
Returns the values of extrapolations of moments of different orders of the distribution moments
}
\usage{
extrapolation.dist(momraw, momdist, ext.rate, momr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{momraw}{
raw moments values
}
  \item{momdist}{
distribution moments values
}
  \item{ext.rate}{
extrapolation rate.
}
  \item{momr}{
values of extrapolation of raw moments to that rate.
}
}
