\name{extrapolation.raw}
\alias{extrapolation.raw}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
extrapolate raw moments across scales
}
\description{
Returns the values of extrapolations of moments of different orders of the raw moments
}
\usage{
extrapolation.raw(momraw, ratio, ba, bba)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{momraw}{
values of raw moments
}
  \item{ratio}{
the ration of x by y.
}
  \item{ba}{
total area we try to extrapolate to.
}
  \item{bba}{
total area that we have informations.
}
}