\name{f}
\alias{f}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
individuals in a given positions
}
\description{
Lista of individuals exist in a given sub-area.
}
\usage{
f(df, xcoor, ycoor, x, ratio)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{
data frame, name of species, position over X and Y axis
}
  \item{xcoor}{
initial X-coordinate inside total area
}
  \item{ycoor}{
initial Y-coordinate inside total area
}
  \item{x}{
sub-section size over Y axis
}
  \item{ratio}{
ratio of X by Y
}
}