\name{fg}
\alias{fg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
list of number individuals of all species
}
\description{
Returns the abundances of species of interest from a given list of individuals.
}
\usage{
fg(g, sppnames)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{g}{
list of all individuals in a sub-area
}
  \item{sppnames}{
list of all species
}
}