\name{fo}
\alias{fo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fo
}
\description{
used as part of method to evaluate the parameter of log-series across scales
}
\usage{
fo(a, n, s)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{
area
}
  \item{n}{
number of individual
}
  \item{s}{
number of species
}
}