\name{fprime}
\alias{fprime}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
first derivative of funtion fo, as a part of the function calculate the parameter of log-series
}
\description{
first derivative of the fo function used for the Newton's method to find the parameter of log-series
}
\usage{
fprime(a, n, s)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{
area
}
  \item{n}{
numbero of individuals
}
  \item{s}{
number of species
}
}
