\name{LLfun}
\alias{LLfun}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Parameters of the best fitting distribution
}
\description{
Returns the parameters of the best fitting distribution with respect to the log-likelihoods.
}
\usage{
LLfun(Y, spp)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Y}{
A matrix consists of x,y and species names respect all individuals in the community
}
  \item{spp}{
list of all species in the community
}
}