\name{LNmom}
\alias{LNmom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Moments of log-normal
}
\description{
Returns moments up to degree k of the log-normal distribution.
}
\usage{
LNmom(mu, sig, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mu}{
mu value in the log-normal distribution
}
  \item{sig}{
sigma value in the log-normal distribution
}
  \item{k}{
the order of moment.
}
}