\name{LSmom}
\alias{LSmom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Moments of the log-series distribution.
}
\description{
Returns moments up to degree k of the log-normal distribution.
}
\usage{
LSmom(x, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
the parameter of log-series
}
  \item{n}{
order of moment
}
}
