\name{M1}
\alias{M1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
M1, theoretical average
}
\description{
Returns average number of individuals (Theoretical)
}
\usage{
M1(ro, c, z, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ro}{
parameter ro in N=ro A
}
  \item{c}{
parameter c in S=c A^z
}
  \item{z}{
parameter z in S=c A^z
}
  \item{A}{
area or scale
}
}