\name{manage_data}
\alias{manage_data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
data across scales, number of individuals of species across scales
}
\description{
returns number of individuals of all species and number of species across scales.
}
\usage{
manage_data(df, spp, n, x, ratio)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{
data frame, name of species, position over X and Y axis
}
  \item{spp}{
list of all species
}
  \item{n}{
number of random sampling over Y axis
}
  \item{x}{
sub-section size over Y axis
}
  \item{ratio}{
ratio of X by Y
}
}