\name{mmoment}
\alias{mmoment}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
first step for finding moment of Possion-Gamma, the values needed to be transformed
}
\description{
the values from this function is used as the first step to calculate the moments of Poisson-Gamma distribution and need to be transformed
}
\usage{
mmoment(al, bet, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{al}{
parameter alpha in the gamma distribution
}
  \item{bet}{
parameter beta in the gamma distribution
}
  \item{k}{
maximum order of moments
}
}
