\name{mo2nd}
\alias{mo2nd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
2nd moment across scales
}
\description{
consider the linear pattern for the 2nd moment across scales.
}
\usage{
mo2nd(res, slope, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
residual of second moment across scales
}
  \item{slope}{
slope of second moment across scales
}
  \item{A}{
area or scale
}
}