\name{moment1}
\alias{moment1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
moment of log-series number of individuals are not log-transformed
}
\description{
return moments of different orders of log-series but in the case where the number of individuals are not log-transformed
}
\usage{
moment1(x, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
parameter of log-series
}
  \item{n}{
maximum moment order
}
}
