\name{mufinderLN}
\alias{mufinderLN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mean finder for the Log-Normal
}
\description{
find the parameter mu in a log-normal
}
\usage{
mufinderLN(res, slope, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
residual of the first raw moment across scales
}
  \item{slope}{
slope of the first raw moment across scales
}
  \item{A}{
area or scale
}
}