\name{mufinderPLN}
\alias{mufinderPLN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mean finder for the Poisson log-normal
}
\description{
returns the parameter of mu in the Poisson log-normal.
}
\usage{
mufinderPLN(res, slope, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
residual of the first moment across scales
}
  \item{slope}{
slope of the first moment across scales
}
  \item{A}{
area or scale
}
}