\name{numsum}
\alias{numsum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
weighted sum
}
\description{
Calculate the weighted sum for given weight and a list
}
\usage{
numsum(n, x, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
maximum degree
}
  \item{x}{
a numeric parameter
}
  \item{k}{
a numeric value for the weight
}
}
