\name{PGmom}
\alias{PGmom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Moments of the Poisson Gamma 
}
\description{
Returns moments up to degree k of the Poisson Gamma distribution.
}
\usage{
PGmom(al, bet, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{al}{
parameter alpha of poisson gamma distribution
}
  \item{bet}{
parameter beta of poisson gamma distribution
}
  \item{k}{
order of moment
}
}