\name{PLNmom}
\alias{PLNmom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Moments of the Poisson log-normal distribution.
}
\description{
Returns moments up to degree k of the Poisson log-normal distribution.
}
\usage{
PLNmom(mu, sig, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mu}{
parameter mu of the distribution
}
  \item{sig}{
parameter sigma of the distribution
}
  \item{k}{
order of moment
}
}