\name{provide_info}
\alias{provide_info}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
provide some info
}
\description{
inside function to print some information from LLfun function.
}
\usage{
provide_info(L)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{L}{
A list of information provided by the LLfun function.
}
}
