\name{ptransform}
\alias{ptransform}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
polylog transformation
}
\description{
Polylog transformation used to calculate moments of log-series where number individuals are log transformed
}
\usage{
ptransform(x, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
parameter for log series
}
  \item{k}{
integer as degree of the polylog function to expand to.
}
}
