\name{r.mom}
\alias{r.mom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
raw moments
}
\description{
calculate raw moments
}
\usage{
r.mom(dat, no.m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
data list of number of individuals of species
}
  \item{no.m}{
maximum order of moments
}
}
