\name{relpwave}
\alias{relpwave}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
wave function
}
\description{
fit wave function and use it for extrapolation
}
\usage{
relpwave(l, t, ba)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{l}{
list of values that is considered to be wave
}
  \item{t}{
the larger scale we need to extrapolate to
}
  \item{ba}{
area or scale 
}
}
