\name{ro.c.z}
\alias{ro.c.z}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
parameter finder
}
\description{
returns theoretical parameter ro, c, and z in N=ro A and S=cA^z
}
\usage{
ro.c.z(No.Sp, No.Id, bba)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{No.Sp}{
number of species across scales
}
  \item{No.Id}{
number of individuals across scales
}
  \item{bba}{
area or scales
}
}
