\name{shrt}
\alias{shrt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Symbolic shortenning
}
\description{
Shorten a polynomial with symbolic calculation
}
\usage{
shrt(u)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
}