\name{sigfinderPLN}
\alias{sigfinderPLN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
standard deviation finder for Poisson Log Normal across scales
}
\description{
returns the parameter sigma in a Poisson log-normal.
}
\usage{
sigfinderPLN(res1, slope1, res2, slope2, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res1}{
residual for the change of means across scales
}
  \item{slope1}{
slope of the change of means across scales
}
  \item{res2}{
residual for the change of moments of second order across scales
}
  \item{slope2}{
slope of the change of moments of second order across scales
}
  \item{A}{
area or scale
}
}
