\name{smom}
\alias{smom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Log-normal moments
}
\description{
moments of log-normal of different orders are calculate here
}
\usage{
smom(mu, sig, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mu}{
mean
}
  \item{sig}{
standard deviation
}
  \item{k}{
maximum order of moment
}
}
