\name{st2k}
\alias{st2k}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sterling number of second kind
}
\description{
Returns the Sterling number of second kind
}
\usage{
st2k(n, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
number of elements of the set
}
  \item{k}{
number of partitions of the set
}
}
