\name{summ}
\alias{summ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
symbolic addition
}
\description{
Adding polynomials symbolically
}
\usage{
summ(u1, u2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u1}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
  \item{u2}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
}
