\name{Tcheb.pol.2nd}
\alias{Tcheb.pol.2nd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Values of Tchebychev polynomials up to certain degree at each bin.
}
\description{
Returns values of the Tchebychev polynomials up to a certain degree and by considering a certain number of bins.
}
\usage{
Tcheb.pol.2nd(N, no.m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
number of bins
}
  \item{no.m}{
maximum order of polynomials
}
}

