\name{tn}
\alias{tn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Thecbychev polynomials recursive formula
}
\description{
calculate recursively the Tchebychev polynomials up to given degree and in different bins
}
\usage{
tn(n, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
maximum degree of polynomial
}
  \item{N}{
maximum number of bins
}
}
