\name{xmul}
\alias{xmul}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
symbolic multiplication
}
\description{
Returns the symbolic multiplications of a polynomial by its variable
}
\usage{
xmul(u)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{
matrix with two rows and second raw consists of 1 to number of columns (represent powers) and first raw is the coefficients
}
}


